Minor cleanups, hopefully final, of the compile-time/runtime thing:
Some time ago, I wrote:

> Consider:

>    TYPE Rng = [0 .. 10];
>    VAR V: Rng := 17;

> Ada explicitly mandates that (in equivalent code), a compiler may
> issue a warning about the guaranteed range fault, but not a compile time
> error, instead generating code that is sure to suffer a runtime error.
> I have always been conflicted about what I think of this, but lean
> towards just preferring the error to occur at compile time.

> Modula3 says nothing about this distinction that I can find.  CM3
> (and probably going back to SRC M3) does it the Ada way.  However,
> it's buggy.  In this case:

>    TYPE Arr = ARRAY [0 .. 1] OF Rng;
>    CONST C = Arr {5, 11};

> where the actual assignment is done at compile time into a static, constant
> data area, it gives the warning (produced by the same compiler code) but
> no error.  Instead, it silently stuffs invalid values into the constant area.

> While reworking array constructors for packed elements, I would prefer to
> fix this by just changing the warning to an error.  It's less work, and
> avoids duplicating the static range check, both code and execution.

> Any opinions?

The responses were 2 to 1 in favor of the current bahavior.  Since
then, I have noticed places in m3core and/or libm3 that depend on this
behavior for code that is word-size-independent, and also places in
ordsets and maybe vararray.  So it seems to be the right way.

However, we have an inconsistent mess about it.  For one thing, the
language definition isn't specific about this, not that I can
interpret.

As for the implementation, this is only done for ordinal types, and
only in assignments and a proper subset of the other cases that
require assignability.  Assignability of procedures also haa s runtime
assignability rule (it can't be nested) that can be sometimes but not
always detected atatically.  References can require a runtime narrow
check, but the only reference value that is constant is NIL, and it
will never fail the check.

In addition, for record, set, and array constructors that are entirely
constant, such assignability violations of components thereof (these
can be detected statically) are only warned at compile time, but never
raise runtime errors when used.

So here is the uniform principle I propose:

In 2.3.1, the definition of assignability of an expression to a
variable requires "the value of e is a member of the type of v, is not
a local procedure, and if it is an array, then it has the same shape
as v."  and, these "generally require runtime checks."

I propose that an implementation is allowed (but not required) to
issue a static error only if both of the following hold:

1) The violation is statically inevitable if the subject code is
   executed, and

2) The code is statically guaranteed to be executed in any main
   program or library that contains it.

When 1) but not 2) holds, the compiler should issue a warning.  The
compiler must generate code that raises a runtime failure whenever 1)
does not hold.

I have changed the packedVars branch to follow this, consistently, I
believe.

Consider this case, in a global scope:

VAR V := ARRAY [ 0 .. 5 ] OF [ 0 .. 10 ] { 25 , .. };

This case meets both criteria.  Morever, portions of the runtime
system are executed before their module's initialization (the only
place the compiler could put a runtime abort) and depend instead on
variables' being statically initialized.  So it is not possible to
always detect the failure at runtime.  So I have made this a static
error.