Fix spurious warning about type of undefined variable being NULL.
This is an interesting bug.  Type.IsEqual considers ErrType.T to
be equal to any type.  This is what one wants when the equality
is expected to be TRUE and a message will be produced otherwise.
When the equality is expected to be FALSE and a meessage is produced
if equal, this is incorrect behavour.

In the subject instance, the type of an undefined variable
produces an unwanted warning, claiming, inaccurately, that
the type is NULL.  Fixed here by replacing IsEqual by plain
pointer equality with Null.T.

Now all calls on Type.IsEqual need to be vetted to see if some
others need this change.  There are quite a few, and many are
comparing to a predefined type.