Remove a warning about cg on-stack types, apparently not relevant.
In CG.m3, ForceStacked:

This seems peculiar.  It converts any Word* type to its
same-sized Int* counterpart on the stack.  ZTypes (in
M3CG.i3), which claims to be the set of arithmetic types
used on the stack, includes Word* types.  Perhaps this is
just an un-updated leftover from Modula2 or 2+.  Perhaps
it all works because no direct stack operators care about
signedness.  Operands of short memory types *are*
sign-extended or zero-extended when loaded, as here. *)