Adapt m3ui for sets not necessarily INTEGER in size.
Matching a 16-element Modula3 set to external C data was assuming
the compiler made all sets at least a whole native word.  In the
packedVars branch, this is no longer so, leading to LOOPHOLE size
mismatch errors when compiling this code.  Now it works for whatever
size the compiler gives the set.