Fix incorrect but undetected array constructor.
The CONST constructor NullAddress16 had a shape mismatch.  The head
compiler has a bug, previously fixed in the packedVars branch, that
does not detect it.  The compiler bug was harmless because the bad
constant is unused in the cm3 distribution.  The packedVars branch
compiler warns that runtime errors would occur if the bad constant
were used, which, using the packedVars compiler, is true.  This fix
suppresses the warning and the latent runtime error.