make-dist: Copy the .objs for all stages. This is required to avoid crashing on AMD64_NT with stack overflow, when running functions with try.
This used to work somehow?

The overall problem here is that when dynamic linking
on NT, dynamically linked symbols must be accessed
through a pointer, the name is prefixed with __imp__.

For functions, the linker will create the function for you:

foo:
 jmp [__imp__foo]

but it cannot do that for data.

it is generally undesirable to micromanage, even in a compiler,
which symbols are imported or not.

There are choices to avoid such micromanagement:
 1 No dynamic linking. Not really acceptable.
 2 No dynamic linking of data. This is done here.
   By putting the data in a few known objs that
   get special treatment. It could be better.
   Another way would really be to dynamically link no data.
   Use accessor functions. That is tempting.
 3 Reference all symbols via a pointer, which
   might just point back to the referencing module.
   Somehow need to manage the defaulting of those.
   I think this is the Unix way. This is tempting,
   but has not been attempted. It might get linker warnings
   but they are ignorable.

Likely advantage of option 3 is pointer identity.

The reason for the stack overflow is we use alloca(m3_jmpbuf_size)
and m3_jmpbuf_size, data, is messed up.
One would think the mess up is it is a pointer to size,
but that didn't seem to be the case.