Config:Fix and reenable parallel backend for NT.common.
The problem was that given:

Foo.m3
Foo.i3

We do
 cl -c foo.m3.c > foo.lst
 cl -c foo.i3.c > foo.lst

In parallel, colliding on foo.lst.

The fix is to instead:
 cl -c foo.m3.c > foo.m3.lst (or even foo.m3.c.lst)
 cl -c foo.i3.c > foo.i3.lst

We only use the last path element but I believe other parts
of builder require that to work.

The direction could also be entirely removed.
It was just to get more deterministic quieter output.