scripts: Fixing AMD64_NT bootstrap Makefile.
For now we have three or four kinds of Makefiles we produce:
 1. NT386 integrated backend.
 2. AMD64_NT C backend.
 3. The rest C backend.
 4. The rest gcc backend?

Perhaps this is just two types. THe NT386 integrated
backend Makefile is the definite outlier.
It is bad, but works, it says like:
 cl *.c -link *.mo *.io

What is bad is that it compiles *.c every time. Not incrementally.
But there aren't many.

The production of the AMD64_NT C backend makefile was getting
confused because the determination of CBackend was in the wrong
order and didn't work.

As such, this diff is mostly just reordering lines and is
difficult to read.

Other than reordering, this also edits commands and amends the logic
slightly:

Before:
CBackend = "c" in LowercaseArgv or "+c" in LowercaseArgv

After:
CBackend = "c" in LowercaseArgv or "+c" in LowercaseArgv or TargetOnlyHasCBackend

The ultimate goal here is, in this part, only support C backend
and produce cmake input but we stumble along for now asis.
Note that cmake does not actually always produce particularly good makefiles.
They run the compiler once per source instead of once.
We might ultimately create one .c file per library anyway.

We might also produce portable Makefiles instead.
With a limited feature set, and not requiring dynamic linking,
it is easy. We'll see. Ultimately this should build the entire system
including dynamic linking.

The link command line is also too long and will be fixed separately.