scripts: NT386 bootstrap Compile C with -MT for multithreaded C runtime instead of -MD for multithreaded dynamically linked C runtime.
The reason is that when using certain older tools, the dynamic
C runtime requires a manifest, and using mt.exe or linker switches.
Which this simple makefile does not know about.

True story, I ran into this tonight.
Because newer toolsets have another harder to solve problem.

The statically linked C runtime does not require this.

This is a tradeoff and there are several alternatives.

This does not affect the "normal" build system, just the
special bootstrap path that only builds cm3 and mklib.