NT bootstrap: Empty LibcCompatC.c.
This file attempts to solve a problem, but it only works
sometimes, and breaks things sometimes.

The problem is taking an existing static m3core.lib/libm3.lib
that contains references to C runtime, generated by compiler
or manual calls to things in .h files, and link that with a newer
compiler/runtime.

The C runtime ABI is unfortunately not generally compatible.
Certainly not going backwards, there are new calls required,
and perhaps not going forward.

This actually broke me yesterday.
An available working "easier" solution is to use boot1.py
and defer any C compilation until you have arrived at your
actual C toolchain. Do not try to mix objects between toolsets.
This is an unfortunate compromise.
In time we might might be able to limit ourselves to a more portable
subset, but that is not super likely, due to things as simple-seeming
but complicated as integer/float conversion.

More likely we should delete this file entirely, rather than
leave it almost empty.