Start writing new documenation/readme.
This documents many things in one small document,
and needs work.

In my opinion the documentation should gloss over
the m3cc and integrated backends and focus on the C backend.

A goal should be to get anyone up and running fast as possible,
with the lowest failure rate, working from "source" (generated)
or binaries.

The .sh and .cmd scripts should be ignored as well.

The .py scripts are even barely useful to a begginer.
 I suppose the primary use case, really, is iterating in on directory,
 in which case cd around and run cm3 devolves to cd once and run cm3.

However the scripts should be documented for people working on the
system. And still, we should try to move toward quake, I think.