m3c:Accomodate sets smaller than a word.
Sets used to be one or more words.

Now they are:
 A single int8 or int16 or int32 or int64, or one or more words
 and not all bits are always used (probably already the case)
i.e. 1, 2, 4, 8, or 4*n or 8*n bytes
not 3 5 6 7 9 15 etc.

The base system includes at least one or two such sets.

(I'm not sure a set can be in an int64 on a 32bit system,
but this is just as well -- array of 2 int32.
The difference hypothetically could be around high level C
code to access; which doesn't exist; which in any case could
always be an array, of possibly size 1, in a struct;
in a struct for copying.)

Perhaps, really, the system-specificity should be removed.
64bit systems should be allowed to use array of int32?
Yielding: 1 or 2 or 4 or 8 or 4*n.

Or just allow n?