cm3,m3middle: Switch to some two level imports.
i.e.
replace FROM Target IMPORT M3BackendMode_t
...M3BackendMode_t...

with:
IMPORT Target;
Target.M3BackendMode_t

It is a tradeoff. The verbosity is not necessarily more readable,
and the second part of the name is globally unique.

This is only done a little bit, in code I added and code
I am likely to change soon (separating cosmetic and semantic
changes).

This is purely cosmetic. There is no semantic change.