config: Let BUILD_DIR from command line always dominate config files.
cm3cfg.common already said:
if not defined("BUILD_DIR")
    readonly BUILD_DIR = ...
end

But NT, Cygwin, Mingw unconditional:
readonly BUILD_DIR = ...

The result of this is that on non-NT systems you can say:

cm3 -DBUILD_DIR=output
but on NT that gives an error about assigning to a readonly variable.

The underlying code is a bit confusing.
First we do the assignment in code.
 I think I put that in.
And we write out quake code to do the assignment,
and before evaluating that we evaluate the config file.

So the sequence is:
  BUILD_DIR=output
  readonly BUILD_DIR=something else
  BUILD_DIR=output error

With this change it should be:
  BUILD_DIR=output
  skip readonly BUILD_DIR=something else
  BUILD_DIR=output no error

Probably it should be possible to say readonly.