Win32: Add some "BITS FOR". Record layout rules recently changed. This broke the important coincidence between windows.h and the Modula-3 rewrite of it in .i3 files. Which perhaps were wrong all along, or skirting the rules. This causes getting a file's size to be misreported, as very large, and then memory allocation to hold it to fail.
In particular
FILETIME = RECORD
 high, low: UINT32
END;

got 8 byte alignment and then catastrophe putting
that in other records, padding got inserted, offsets
changed, etc.

This goes to show perhaps that rewriting .h files
in .i3 is a bad idea, if you do not have control of those .h files.

If we owned the .h we would only have INTEGER and LONGINT
and nothing ever smaller. And we'd have even number of runs
of INTEGER, so there'd never be any implicit padding for alignment.

And we'd have only what we use, and we might assert correctness
at startup.

There could very well be impact to other platforms esp. 64bit ones.
Within the Critical Mass system, impact is mitigated by reducing
such header rewrites, but they are not eliminated, there could still
be trouble?