Current MIPS64/Linux and ARM32/Linux (armhf) seem to support atomics over 1, 2, 4, 8 so remove prohibition against similar but not same: ARMEL_LINUX, MIPS64_OPENBSD, MIPS64EL_OPENBSD.
PPC and Sparc32 remains as purportedly missing atomic double.
mips32 probably too. Investigate later/never.

It is possible OpenBSD's compiler has less support.
And when ARMEL_LINUX was introduced it was like v5 or v6.

These otherwise are a barrier to portable distribution.
Perhaps we do need a way to tunnel through this conditionality though.