Rework widechar size initialization.
Unfortunately, widechar is 16 or 32bits, depending
on command line and config. This seem dubious but this
PR does not change it.

The initialization worked by evaluating cm3.cfg a second time,
by virtue of MxConfig.Get(), ignoring the command line.

Evaluating cm3.cfg a second time, while ignoring command line
is a bad idea. It issues ignored errors when you try things like
cm3 -DROOT or -DTARGET.

This PR does have the side effect of cm3 command line working
to set the size, via -DUnicode_WIDECHAR

But the real point is to not evaluate cm3.cfg twice,
in particular, the second evaluation ignores the command line.

Previously widechar initialization was in cm3 and poked locals
in m3linker and m3front.

Original author did not realize that m3middle's role is to be shared
by various. And Target seems like the main interface in m3middle.
I just add stuff to it.

So this PR moves most of the code to m3middle.

Along with renaming everything.

There is now:

  const Target.WideChar16Max = 16_ffff
  const Target.WideChar32Max = 16_10ffff
  Target.WideCharMax() WideChar16Max or WideChar32Max
  Target.WideCharSize() 16 or 32
  Target.SetWideChar16() set by cm3 in command line parsing, or from config
  Target.SetWideChar32() set by cm3 in command line parsing, or from config