m3c: Allow invalid IR: direct calls with insufficient parameters
m3front is willing to issue an unconditional abort
in the middle of start_call/call.
Upon which, it stops popping parameters.

This leads to errors from the C compiler, for calls
with insufficient parameters.

Test p268 does this.

When we see this, evaluate whatever parameters we get,
possibly before the abort, maybe that has visible
side effects, but skip the call.

Now p268 compiles.

Now that indirect calls in m3c are presently untyped.
This should apply there as well, but the C compiler
would already accept what we had.

This assumes no continue after abort, which is a little debatable.
But we don't have enough parameters from m3front so we cannot very
well make the call. m3front ignores that you can continue from
faults in a debugger.

We should probably make this conditional on if we saw all the parameters.

We should probably expect better IR from m3front also.