cm3: Remove commented out code to default target to host properties:
 C_COMPILER: meant to be SUN or GNU (i.e. on Solaris targets)
 LINKER: MS or GNU (i.e. I386_NT or I386_CYGWIN)
   Affects linker switch format.
 WINDOWS_LIBRARY: MS (win32) or X
 THREAD_LIBRARY: MS (Win32) or pthread
   Cygwin was actually configured to use MS but either make sense.
 BACKEND_MODE: 0-3, C
 HOST_GNU_MAKE: make vs. gmake etc.

The code was never active here and is little known.
The options, which remain a useful and implemented
 factoring of config, can be set from command line with -D
  or config file or maybe environment (hopefully with CM3_ prefix).

Defaulting them to host would introduce somewhat undesirable host variation.
  Granted, a host variation that might be sticky anyway (i.e.
  these correlate strongly with Win32 vs. Posix, a factoring we might
  be stuck with forever).

The idea being to not factor on coarse grained Win32 vs. Posix,
but pieces. Because you can have pthreads on Win32, for example.

If this default target=host is desired, refactor into C.
That will be done for the non-commented out ones.