version.quake: Trim more host properties intended to be target defaults, but that can be set from config or command line and are most likely never used from here:   GNUPlatform   CCompiler   Linker   ThreadLibrary   WindowLibrary   GNUMake   BackendMode   TargetNaming   NamingConventions
Using them would cause changes in backend output and is not particularly worthwhile.
If they are really needed, we can feed them into generated C instead of .m3/.i3
so they do not affect backend output, but that still breaks portable redistributable.

Another way to do is configure defaults when you compile cm3, such as autoconf/cmake do.
That can generate .c or .i3/.m3, but just do it later.

We could imagine a two phase build where we build cm3 w/o default, then run this
quake to update cm3 w/ defaults. But I don't think anyone ever really used this code
and we can restore it then.

Or just write a text file next to the executable, i.e. the config.
It depends on your taste for quantity of files and text parsing and startup I/O.
i.e. the other extreme is bundling config into binaries, like gnu ld spec files,
which would be nice, if users really tend not to need to edit them (which the escape
hatch of dump specs, and edit or use the path if it exists).