config and scripts: Remove the special support for supressing symbols, that I added, for the sake of bouncing between C toolsets. Only for NT (which possibly has the worst compat in this area, of mixing C toolset versions). It surely worked, but bouncing toolset version is probably too much hassle to be worth it. The reliable approach that involves no toolset mixing is to bootstrap from generated C which is totally viable now (I need to double check NT386, but even so).
If someone again starts the exercise of mixing .objs among toolsets,
or creating packages that interoperate across toolsets, this
could be brought back.

It really is a nice idea to create a Modula-3 NT redistributable
toolset, that is not tied to a specific Visual C++ compiler/linker/runtime,
but it might be too much trouble.

Maybe better of to reduce static linking, which is where these
problems creep in.