Solaris: pthread_atfork handlers run with signals deferred/disable. (#258)
Solaris: pthread_atfork handlers run with signals deferred/disabled,
leading to deadlock between fork and collect.
It may be desirable to have the collector "back off" more,
releasing locks. However instead here "for now" instead
run the fork handlers outside of fork on Solaris where
signals work, so forker will suspend.
The usual suspect better approaches of cooperative suspend and spawn instead of fork + exec are not solutions on their own.
The fork handlers are for anyone in the process that forks without exec.
The scenario of anyone doing that, on Solaris, not going through m3core, remains subject to lock and state corruption.