Solaris: Use sigsetjmp instead of _setjmp. (#248)
This is to fix building Solaris C++.
This is a nice portable idea.
But, sigsetjmp should be paired with siglongjmp,
which is ok, but sigsetjmp is not portably linkable,
i.e. it fails with Linux/m3cc.

So do it for Solaris only.

Solaris/m3cc not tested but probably works, just Solaris/m3c.