Remove OSF/1 and Ultrix socket hacks. (#281)
Specifically for certain socket errors, calls are made in an attempt to get the correct error.
Ultrix is out of date 20+ years.
OSF/1 is out of date 10+ years.
This does not remove support for them per se, but just a bit of socket code specifically for them.
If someone brings up these machines (even emulated) and gets everything else working (should not be difficult esp. for OSF/1) and provides access to the machine and preferably can demonstrate these bugs (or not), we will gladly bring this code back.