m3-sys: Consolidate strings for CG type names. (#290)
PROCEDURE Target.TypeImage() => CONST Target.TypeNames[]
 CONST CG.TypeName[] =>  CONST Target.TypeNamesFixedWidth[]
 CONST Target.TypeNames[] => CONST Target.TypeNamesDotted[]

Arguably should be VAR or PROCEDURE but the main
point is they are all Target.TypeNames. Two were
already CONST so stick with that.

Arguably do not need three forms, but the point
is consolidation not behavioral change, and exact
printed strings are behavior.

This is cleanup. Keep the duplication, but put it all near each other.