m3core: Refactor GetPC into a GET_PC and usually function GetPC. The reason for this is speculation that we should compute HOST amidst the same code. If we don't just use uname.
A full static host is fairly tempting, except, that on Debian
uname.machine factors in the endian vagaries.
Which I don't want to figure out, or worse, name.

Therefore HOST should be computed statically in many cases, but not all.
i.e. when GetPC had good clear support, or mostly (sometimes word
size is ambiguous), but not in the cases where endian can vary.
i.e. endian is unambiguous for x86, amd64, mac/ppc, ia64, sparc, arm64, riscv, m68k, hppa, alpha.
Endian is ambiguous for mips, ppc, arm.
Support is missing for m68k.
Also anything we don't know we will fallback to uname.
We might fallback to uname for a lot.
Also the way NetBSD is handled here, nicely, we should fallback to uname.
Or maybe just uname mostly, but this PR is ok.