Fix long-standing bug in New, uncovered by better type checking in ArrayExpr.
This is a bit complicated.

The bug in New.m3 occurred when allocating an object with field
initializatons.  It was checking type assignability of an
initializing expression against the wrong object field.  This
was originally covered up by incomplete checking when the
expression was an array constructor.  Recent completion of
this checking uncovered the bug, with a symptom of false
error messages, which then were not recovered from, resulting
in massive internal CG failures.

Note: The code being compiled does not have type assignabilty
errors.  The newly emitted messages were bogus, as result of
the field mismatch.

It is possible that the same error message, when produced
legitimately, still fails to recover.