m3c: Fix typo declaring pointers and resulting fallout.
There are typoed recursive types in the system:
  TYPE FOO = REF FOO;

We reluctantly accept these since the rest of the system does.
When we see this, we attempt to break the loop
by just treating it as TYPE FOO = ADDRESS.

Luckily they are always self-referential.
We do not have:
 TYPE FOO = REF BAR;
      BAR = REF FOO;

which would take more work to detect.
We do in some sense detect these, by erroring, due to
a lack of forward progress, described below.

The detection was supposed be like:
  if type.typeid == type.points_to_typeid
    fallback to address

but was accidentally:
  if type.typeid == type.typeid

which is always true.

This was in type.canBeDefined, not type.define.

This lead to calling type.define prematurely.

This lead to type.define falling back and typedefing
types to ADDRESS instead of something strong (pointer
is correct, but pointer to what?).

type.define really should have errored in this case, and now it does.

This all worked ok. Just that debugging might suffer.

However, if you go and concatenate files, this occurs only sometimes.

So we have a mix of:
  struct foo;
  typedef foo* pfoo; // desired
and
 typedef void* pfoo; // bogus fallback

which is fortunately an error in C.

Once you fix this (typeid == typeid vs. typeid == points_to_typeid),
you then run into slight misimplementation of open arrays and forward declaration.

Open arrays require a definition or at least forward declaration
of their constituent elements. So they can declare a pointer to them.

The implementation of forward declare ignored if a type can
be forward declared. For example records/structs can be forward
declared and pointers cannot be.

As a result, you could get an open array with a pointer to a
completely undeclared type. This is also an error.

The fix for this is to call canBeForwardDeclared before
calling forwardDeclare and setting the type's state to forward declared.

Calling forwardDeclare does nothing for types that cannot be
forward declared, but setting the state to forward declared, enables
the open array declaration to proceed.

It should be noted/reminded that the way all the type declarations works:
 - Build up in memory a representation of all of m3front's type declarations.
 - Iterate through the types repeatedly, forward declaring and defining
   what/when can be. Loop until no forward progress is made or until all
   types are defined. If no forward progress, error.
 - m3front is not presumed to output in any particular order.