m3c: Tolerate duplicate typeids.
Subranges have colliding typeids. m3front bug?
For example:
 libm3\Formatter.m3:
    Int   = REF INTEGER;
    ints: ARRAY [-256..256] OF Int;
    declare_subrange typeid:T69A2A904 domain_type:T195C2A74 min:-256 max:256 bit_size:16
 libm3\Sx.m3
    MinBoxedInt   = -100;
    MaxBoxedInt   = 100;
    BoxedInts  := ARRAY [MinBoxedInt .. MaxBoxedInt] OF REF INTEGER {NIL, ..};
    declare_subrange typeid:T69A2A904 domain_type:T195C2A74 min:-100 max:100 bit_size:8

Partial workaround it here by appending the size to the type text.
We could further workaround by deferring the typedef until use, as in this
case, neither type is used. Just defined.

This shows up if you concatenate all the files that make up cm3,
which is a goal for easier redist and build.