x-opengl: Fix calling convention of callbacks.
This fixes compiling x-opengl using the C backend.

Generally calling conventions, having more than one, is an anomoly.
It is only for NT/x86.

However when using the C backend, m3middle/m3front/backend deliberately
do not know this.

So for example, I386_NT is the same as ARM32_NT.

So all code needs to be written as if it is targeting NT/x86,
regarding calling conventions.

The fix is trivial and in good style.
Instead of casting to PROCEDURE(), cast to GLu.GLUtessAnyProc.