Remove 2 overzealous assertions involving multi-D open array constructors.
Makes csv and csvalign compile.

The failing case was an array constructor with at last two open dimensions
but empty between the braces.  The assertions expected no dynamic levels
when compiling a CONST array, but in this case, the deeper open dimensions
are such levels.  But they are irrelevant, because compiling the array
value will not reach them.

In all the other cases, the assertions are both true and needed, but
other assertions will assure this in other ways.