m3front: Pass on result_qid to import_procedure. m3c: Pass on result_qid.
This is a bit of a temporary failure.
The names are coming through with a single item,
"int" instead of "Ctypes__int".

We will maybe want to call GlobalName in m3front like we do for
declare_typename and pass on a TEXT. (usually nil).

Alternatively, we can change the few we care about on the .i3 side
and have m3c only pay attention to ones with module # 0.

As well we can mechanize "what we care about" as extern-only.

There are likely more options, such as a narrow form of GlobalName,
that returns the module M3ID. Or providing more typedefs
like Uuio__int, etc.