m3front: Provide a qid to backend for non-named types. (#424)
That is, tell backend "INTEGER" and not just int32/int64.
Original attempt was to wrap all types in NamedType.
That still seems reasonable.
But it did not immediately work, I did not debug, and the approach here is to add M3ID name to every Type.info.
This is wasteful but probably ok. Wrapping in NamedType is quite disruptive perhaps.. or should work.

This way we can declare C code as taking/returning INTEGER and have it match m3c output (once m3c processes this data more).

Perhaps at this point we should just change qid=m3idpair to typename=m3id, really.
That is, remove the module: M3ID and have NamedType do a concat and M3ID.Add.
However that has cost, and with most backends, no benefit.