m3front: Formal: Add original_type that takes the (#446)
first non-nil value of type, and then is not
reduced/replaced upon calls to .Check/.CheckInfo.
i.e. it retains NamedType.

Return original_type in Formal.Info.type, so that
splitting and reforming a Formal, as is done, retains the NamedType.

This will lead to duplicate work to reduce the same
types multiple times but that should be ok, not terrible,
and ultimately required.