m3front: Variable: Add notion of original_type. (#452)
i.e. the value before Check lowers it
i.e. it remains NamedType (if it is NamedType)
and Typename "works" (not a hash).

Starting using it for parameters, though put off
handling indirect parameters briefly.

We should probably introduce TypedValue.T
as a common base of Formal.T and Variable.T, or push
more functionality up to the base Value.T.