m3-sys: Add typename to declare_formal. (#453)
This might not be needed but it is the right thing.
In particular, this is for function pointer types only,
and if any of those occur where it matters, they can be
replaced entirely instead via their pieces at low implementation
cost, i.e. via the toplevel. We'll see.

Add printing of declare_formal's typename in M3C.
It is never currently passed from m3front.