m3front: Remove my recent addition of original_type and OriginalTypeOf. (#520)
They have been replaced by selectively not lowering type.
To retain NamedType.
This approach might not scale though. There are many other places
in m3front where original types are lost, relying
on backends to aggressively cast to a generic pointer type,
and to have structs that are just arrays of bytes.