m3c / m3core: Repair integer types.
m3front issues:
 declare_typename INTEGER size_t  very questionable
 declare_typename INTEGER ssize_t ok

To cope with this, maybe more than necessary:
 - m3core.h: provide the various unsigned, unsigned_long, unsigned_int, etc. types.
   When combined with m3c output and m3core.h comes first,
   this will lock them in as unsigned as they probably should be.
   It is in their names, right?

   This is only needed in as much as these types occur in
   prototypes in hand written C, so this is probably overkill,
   which is why I said "and then some".

 - m3c: In declare_typename, stop replacing types
   with typenames; so for example INT64 will not change to size_t.
   Typenames remain for parameters, for which the aggressive
   typename work was initially intended, but not on types,
   where I was probably being too clever, not then
   being used where there are no typenames.

This addresses the error resulting from the combination of:

 ADDRESS Cstddef_I3(Cstddef__ptrdiff_t)
 ADDRESS Cstddef_I3(Cstddef__size_t)

which is obviously wrong. These typename-less prototypes
come from special case code in m3front. Another fix
would be to give them typenames for consistency.
Which is to say, all parameters should have typenames.
The types of variables matter less, because there is
no "combination", and their uses are full of casts.

In working on this, it struck me that VMS needs work.
It violates the presumed size matches among pointer, size_t, intptr_t, INTEGER.
Like m3core/src/C/VMS will be needed if we ever really run on VMS.