m3c: Workaround unsigned type problems.
We get
  declare_typename(INTEGER, "ssize_t") ok
  declare_typename(INTEGER, "size_t")  wrong

This is just wrong.
The same type cannot be both signed and unsigned.
Typenames become the names for the underlying type,
and due to typenames occuring in inconsistent order,
this leads to errors, when you combine m3c output
with hand written C (m3core.h).

The workaround is to have m3c try to know
all the incorrectly specified types, and when seen,
do not replace the underlying type's name with them.

Another fix might be to ensure all parameters (locals? globals?)
have typenames, and then the names of types without context
would never be used.

We should consider eliminating unsigned types from use,
esp. uint64 on 64bit, and uint32 and uint64 on 32bit.

Given C code that is usually just pass through, getting
signedness correct doesn't really matter. It has been
and remains wrong all along. We only notice when we
try to make the declarations visible together.
Modula-3 continues to interpret size_t as signed,
unless you think to use size_t with Word.T.
It is a mess.